Add post-install command execution functionality
Added functionality to execute a command after installation in the
`InstallerWindow` class. Introduced a new property `ExecuteAfterInstall`
in the `InstallManifest` class to store the command. Updated the
`ComposerWindow.xaml` to include a new `TextBox` for inputting the
command. Modified the `ComposerWindow` class to handle the new
`ExecuteAfterInstall` property by retrieving the text from the new
`TextBox` and setting it in the `InstallManifest`.